fix broken hook due lib change
