+++ src/components/UserVestingInfoModal/index.ts
@@ -1,2 +1,1 @@
 export { default } from './UserVestingInfoModal';
-export { default as useVestingContract } from './useVestingContract';
